refactor(gui): rename Parser to ArgsParser for clarity
- Rename Parser.hpp/cpp to ArgsParser.hpp/cpp to better reflect
  the specific purpose of parsing command line arguments
- Update include statements in main.cpp to use new filename
- No functional changes, only improved naming convention
